requirements.txt
Django==1.11.6
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-HTTPAuth==3.2.3
PyJWT==1.5.3
Flask-JWT==0.3.2
mysqlclient==1.3.12
